change link colors

+++ dentalapp/models.py
@@ -23,7 +23,7 @@ class Supplier(models.Model):
 
 class Customer(models.Model):
     customer_name=models.CharField(max_length=40, default=None)
-    contact_number=models.IntegerField(default=None)
+    contact_number=models.BigIntegerField(default=None)
     
     def __str__(self):
         return self.customer_name
